Removed obsolete files. Library setup tools now reside in the setup directory. Various bug fixed. Examples updated according to the library changes.
git-svn-id: https://boost-log.svn.sourceforge.net/svnroot/boost-log/branches/bleeding-edge@758 ab6a8de6-4e00-4c3d-a860-8899c17d353f

+++ boost/log/sinks/unbounded_ordering_queue.hpp
@@ -204,7 +204,7 @@ protected:
         {
             const boost::log::aux::timestamp now = boost::log::aux::get_timestamp();
             enqueued_record const& elem = m_queue.top();
-            if ((now - elem.m_timestamp).milliseconds() >= m_ordering_window)
+            if (static_cast< uint64_t >((now - elem.m_timestamp).milliseconds()) >= m_ordering_window)
             {
                 // We got a new element
                 rec = elem.m_record;
